Alter structure of project articles
The anchor tag has been moved inside the article tag for each project.
This is so the HTML makes more sense semantically, and results in just
the picture of the project being clickable.  Having only the images
being clickable in this way also leads to some nice separation between
the projects, as now the projects section isn't just one continuous
clickable block.